% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_sparse.R
\name{cor_sparse}
\alias{cor_sparse}
\title{Find the Pearson correlation of a sparse matrix efficiently}
\source{
This code comes from mike on this Stack Overflow answer:
\url{http://stackoverflow.com/a/9626089/712603}.
}
\usage{
cor_sparse(x)
}
\arguments{
\item{x}{A matrix, potentially a sparse matrix such as a "dgTMatrix" object}
}
\description{
Find the Pearson correlation of a sparse matrix.
For large sparse matrix this is more efficient in time and memory than
\code{cor(as.matrix(x))}. Note that it does not currently work on
simple_triplet_matrix objects.
}
