% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_scale.R
\name{multi_scale}
\alias{multi_scale}
\title{Multidimensional Scaling of documents separated by a distance measure.}
\usage{
multi_scale(tbl, item1, item2, value, k = 2)
}
\arguments{
\item{tbl}{Table obtained by running a pairwise distance method \code{pairwise_delta} or \code{pairwise_dist}}

\item{item1}{first item}

\item{item2}{second item}

\item{value}{Value}

\item{k}{number of dimensions, defaults to 2}
}
\value{
Returns a function that takes at least four arguments:
  \item{item}{Column to store documents or items separated by various distances as used prior to calling \code{multi_scale()}}
  \item{V1}{First Dimension}
  \item{V2}{Second Dimension}
  \item{...}{Other Dimensions as specified by k's value}
}
\description{
Given a tbl or data frame of pairwise distances between documents, scale each document
to a *k* dimensional space that ensures the distance between all documents is maintained.
**NOTE:** Doesnt work when the pairwise distance tbl is formed using \code{upper = FALSE}.
}
\examples{

library(janeaustenr)
library(dplyr)
library(tidyr)
library(tidytext)

# closest documents in terms of 1000 most frequent words
austen_delta <- austen_books() \%>\%
  unnest_tokens(word, text) \%>\%
  count(book, word) \%>\%
  pairwise_delta(book, word, n)

austen_delta

austen_delta \%>\%
  multi_scale(item1, item2, delta)

}
