% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_cor.R
\name{pairwise_cor}
\alias{pairwise_cor}
\alias{pairwise_cor_}
\title{Correlations of pairs of items}
\usage{
pairwise_cor(tbl, group, item, value, method = c("pearson", "kendall",
  "spearman"), use = "everything", ...)

pairwise_cor_(tbl, group, item, value, method = c("pearson", "kendall",
  "spearman"), use = "everything", ...)
}
\arguments{
\item{tbl}{Table}

\item{group}{Group within which to look for correlations}

\item{item}{Item to compare; will end up in \code{item1} and
\code{item2} columns}

\item{value}{Value}

\item{method}{Correlation method}

\item{use}{Character string specifying the behavior of correlations
with missing values; passed on to \code{cor}}

\item{...}{Extra arguments passed on to \code{squarely},
such as \code{diag} and \code{upper}}
}
\description{
Correlations of pairs of items
}
\examples{

library(dplyr)
library(tidytext)

data(AssociatedPress, package = "topicmodels")
ap_terms <- tidy(AssociatedPress) \%>\%
  group_by(term) \%>\%
  filter(n() >= 50)

ap_terms \%>\%
  pairwise_cor(document, term, count, sort = TRUE, upper = FALSE)

}

