% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_cor.R
\name{pairwise_cor}
\alias{pairwise_cor}
\alias{pairwise_cor_}
\title{Correlations of pairs of items}
\usage{
pairwise_cor(tbl, item, feature, value, method = c("pearson", "kendall",
  "spearman"), use = "everything", ...)

pairwise_cor_(tbl, item, feature, value, method = c("pearson", "kendall",
  "spearman"), use = "everything", ...)
}
\arguments{
\item{tbl}{Table}

\item{item}{Item to compare; will end up in \code{item1} and
\code{item2} columns}

\item{feature}{Column describing the feature that links one item to others}

\item{value}{Value}

\item{method}{Correlation method}

\item{use}{Character string specifying the behavior of correlations
with missing values; passed on to \code{cor}}

\item{...}{Extra arguments passed on to \code{squarely},
such as \code{diag} and \code{upper}}
}
\description{
Correlations of pairs of items
}
\examples{

library(dplyr)
library(gapminder)

gapminder \%>\%
  pairwise_cor(country, year, lifeExp)

gapminder \%>\%
  pairwise_cor(country, year, lifeExp, sort = TRUE)

}

