% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_dist.R
\name{pairwise_dist}
\alias{pairwise_dist}
\alias{pairwise_dist_}
\title{Distances of pairs of items}
\usage{
pairwise_dist(tbl, group, item, value, method = "euclidean", ...)

pairwise_dist_(tbl, group, item, value, method = "euclidean", ...)
}
\arguments{
\item{tbl}{Table}

\item{group}{Group within which to look for correlations}

\item{item}{Item to compare; will end up in \code{item1} and
\code{item2} columns}

\item{value}{Value}

\item{method}{Distance measure to be used; see \code{\link{dist}}}

\item{...}{Extra arguments passed on to \code{\link{squarely}},
such as \code{diag} and \code{upper}}
}
\description{
Distances of pairs of items
}

