% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widely_svd.R
\name{widely_svd}
\alias{widely_svd}
\alias{widely_svd_}
\title{Turn into a wide matrix, perform SVD, return to tidy form}
\usage{
widely_svd(tbl, item, feature, value, rank = NULL)

widely_svd_(tbl, item, feature, value, rank = NULL)
}
\arguments{
\item{tbl}{Table}

\item{item}{Item to perform dimensionality reduction on; will end up in \code{item} column}

\item{feature}{Column describing the feature that links one item to others.}

\item{value}{Value}

\item{rank}{Optional; the maximum dimensionality of the data. Recommended for matrices
with many features.}
}
\description{
This is useful for dimensionality reduction. Work in progress.
}
